Track dependencies from beginning to end
The previous version of calculate_dependencies tracked dependencies
based on components, so there was no way to figure out which tasks were
required to meet the dependencies of initial targets.  To try and do
only the minimal work required (trying to consider weird projects where
multiple logical components may use a single checkout), the initial work
is seeded with all requested tasks.  As dependencies are tracked,
they're tracked based on the speciic dependent task, not the dependent
component.